Fixed bug that made ball go left on start
Bug happened because location values were integers, and any small change made them round down.
I replaced x and y values in all Items to doubles and casted the values to int when drawing them on the window.

+++ Item.java
@@ -2,8 +2,8 @@
 
 public abstract class Item
 {
-    public int _x;
-    public int _y;
+    public double _x;
+    public double _y;
 
     
     public Item()
@@ -11,7 +11,7 @@ public abstract class Item
         _x = 0;
         _y = 0;
     }
-    public Item(int x, int y)
+    public Item(double x, double y)
     {
         _x = x;
         _y = y;
